Use flow-go VM to unify the execution (#42)
* Migration to flow-go types
* SDK types for interface, Flow types for internal
* Migrate to VM validating signatures
* Runtime keys
* Sequence key errors
* Cadence bump
* Adapt to flow-go changes
* Review fixes
* Bump to flow-go master with VM changes merged
*Adapt changes to flow block model

+++ storage/badger/keys_test.go
@@ -42,7 +42,7 @@ func TestKeyOrdering(t *testing.T) {
 		for _, num := range nums {
 			for i := 0; i < 3; i++ {
 				for j := 0; j < 3; j++ {
-					keys = append(keys, eventKey(num, i, j, "foo"))
+					keys = append(keys, eventKey(num, uint32(i), uint32(j), "foo"))
 				}
 			}
 		}
